added tests, some code moved to separate module, active negotiation pop-up removed
