+++ bootstrap-js/passive-negotiation.html
@@ -17,10 +17,8 @@
     <title>Showing the cards in a web page</title>
     <link rel="stylesheet" href="ts.css">
     <script type="text/javascript" src="inc/web3.min.js"></script>
-<!--    <script type="text/javascript" src="inc/require.js"></script>-->
     <script type="text/javascript" src="inc/ethers-5.0.umd.min.js"></script>
-<!--    <script type="module" src="tokenscript-bootstrap.js" ></script>-->
-    <script type="text/javascript" src="tokenscript-bootstrap.js" ></script>
+    <script type="module" src="tokenscript-bootstrap.js" ></script>
     <script type="text/javascript">
         // the following event is triggered every time when a token is negotiated
         // and that is why an event listener is used in place of a promise;
