+++ bootstrap-js/ts.scss
@@ -131,7 +131,7 @@
     box-shadow: inset 0 -3.25em 0 0 #eee;
   }
 }
-
+/*
 #ts_select_token {
   width: 100%;
   display: flex;
@@ -200,6 +200,7 @@
     }
   }
 }
+ */
